sharing_constructs/pyproject.toml
[tool.poetry]
authors = ["Andre Sionek <andre@sionek.com>"]
name = "best_practices"
description = "Sharing Constructs"
keywords = ['cdk', 'iac', 'infrastructure', 'devops', 'python', 'backend', "platform"] # noqa
readme = "README.md"
homepage = "https://sionek.com/"
version = "0.0.0"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.99.0"
cdk-constructs = {git = "https://github.com/Real-Life-Infrastructure-as-Code/cdk-constructs.git", rev = "latest"}
